refactored health field renderer into more generic integer field renderer
